+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Changelog
 - CLI: added progress bar.
 - CLI: fixed bug that prevented from using "login" subcommand.
 - CLI: skip already downloaded/uploaded files.
+- added logging to ``todus.client`` and make the CLI write logs to ``~/.todus/log.txt`` file.
 
 0.1.0
 -----